Move changes to main vbet
