+++ packages/vbet/.vscode/launch.json
@@ -59,11 +59,20 @@
       "args": [
         "${input:HUC}",
         "UPDATED_TESTING",
-        "FLOWLINES={env:DATA_ROOT}/rs_context/${input:HUC}/hydrology/hydrology.gpkg/network,FLOW_AREAS={env:DATA_ROOT}/rs_context/${input:HUC}/hydrology/NHDArea.shp,SLOPE_RASTER={env:DATA_ROOT}/taudem/${input:HUC}/outputs/gdal_slope.tif,HAND_RASTER={env:DATA_ROOT}/taudem/${input:HUC}/outputs/HAND.tif,TWI_RASTER={env:DATA_ROOT}/taudem/${input:HUC}/outputs/twi.tif,CATCHMENTS={env:DATA_ROOT}/rs_context/${input:HUC}/hydrology/NHDPlusCatchment.shp,CHANNEL_AREA_POLYGONS={env:DATA_ROOT}/channel_area/${input:HUC}/outputs/channel_area.gpkg/channel_area,HILLSHADE={env:DATA_ROOT}/rs_context/${input:HUC}/topography/dem_hillshade.tif,DEM={env:DATA_ROOT}/rs_context/${input:HUC}/topography/dem.tif",
+        "{env:DATA_ROOT}/rs_context/${input:HUC}/hydrology/hydrology.gpkg/network",
+        "{env:DATA_ROOT}/rs_context/${input:HUC}/topography/dem.tif",
+        "{env:DATA_ROOT}/taudem/${input:HUC}/outputs/gdal_slope.tif",
+        "{env:DATA_ROOT}/rs_context/${input:HUC}/hydrology/NHDPlusCatchment.shp",
+        "{env:DATA_ROOT}/channel_area/${input:HUC}/outputs/channel_area.gpkg/channel_area",
         "{env:DATA_ROOT}/rs_context/${input:HUC}/hydrology/nhd_data.sqlite/NHDPlusFlowlineVAA",
         "{env:DATA_ROOT}/vbet/${input:HUC}",
-        "--reach_codes",
-        "33400,46000,46003,46006,46007,55800",
+        "--level_paths","${input:level_paths}",
+        "--pitfill","{env:DATA_ROOT}/taudem/${input:HUC}/intermediates/pitfill.tif",
+        "--dinfflowdir_ang","{env:DATA_ROOT}/taudem/${input:HUC}/intermediates/dinfflowdir_ang.tif",
+        "--dinfflowdir_slp","{env:DATA_ROOT}/taudem/${input:HUC}/outputs/dinfflowdir_slp.tif",
+        "--twi_raster","{env:DATA_ROOT}/taudem/${input:HUC}/outputs/twi.tif",
+        //"FLOWLINES={env:DATA_ROOT}/rs_context/${input:HUC}/hydrology/hydrology.gpkg/network,FLOW_AREAS={env:DATA_ROOT}/rs_context/${input:HUC}/hydrology/NHDArea.shp,SLOPE_RASTER={env:DATA_ROOT}/taudem/${input:HUC}/outputs/gdal_slope.tif,HAND_RASTER={env:DATA_ROOT}/taudem/${input:HUC}/outputs/HAND.tif,TWI_RASTER={env:DATA_ROOT}/taudem/${input:HUC}/outputs/twi.tif,CATCHMENTS={env:DATA_ROOT}/rs_context/${input:HUC}/hydrology/NHDPlusCatchment.shp,CHANNEL_AREA_POLYGONS={env:DATA_ROOT}/channel_area/${input:HUC}/outputs/channel_area.gpkg/channel_area,HILLSHADE={env:DATA_ROOT}/rs_context/${input:HUC}/topography/dem_hillshade.tif,DEM={env:DATA_ROOT}/rs_context/${input:HUC}/topography/dem.tif",
+        "--reach_codes", "33400,46000,46003,46006,46007,55800",
         "--flowline_type", "NHD",
         // "--epsg", "4326",
         "--meta",
