+++ packages/vbet/vbet/vbet.py
@@ -1,49 +1,45 @@
-# Name:     Valley Bottom
+# Name:     Vbet Centerlines (raster)
 #
-# Purpose:  Perform initial VBET analysis that can be used by the BRAT conservation
-#           module
+# Purpose:  Find the vbet centerlines per level path via rasters
 #
-# Author:   Matt Reimer, Kelly Whitehead
 #
-# Date:     April 9, 2021
+# Author:   Kelly Whitehead
 #
-# Vectorize polygons from raster
-# https://gis.stackexchange.com/questions/187877/how-to-polygonize-raster-to-shapely-polygons
+# Date:     Apr 11, 2022
 # -------------------------------------------------------------------------------
-import argparse
+
 import os
 import sys
+import argparse
 import traceback
-import json
 import time
 from typing import List, Dict
-# Leave OSGEO import alone. It is necessary even if it looks unused
-from osgeo import gdal, ogr
+
+import gdal
+from osgeo import ogr
 import rasterio
-from rasterio.features import shapes
-import rasterio.mask
+from rasterio import shutil
 import numpy as np
-from shapely.geometry import MultiPolygon
+from scipy.ndimage import label, generate_binary_structure, binary_closing
+
+from rscommons import RSProject, RSLayer, ModelConfig, ProgressBar, Logger, GeopackageLayer, dotenv, initGDALOGRErrors
+from rscommons.vector_ops import copy_feature_class, polygonize, get_endpoints
+from rscommons.util import safe_makedirs, parse_metadata, pretty_duration
+from rscommons.hand import run_subprocess
+from rscommons.vbet_network import copy_vaa_attributes, join_attributes, create_stream_size_zones, get_channel_level_path
 from rscommons.classes.rs_project import RSMeta, RSMetaTypes
 
-from rscommons.util import safe_makedirs, parse_metadata, pretty_duration
-from rscommons import RSProject, RSLayer, ModelConfig, ProgressBar, Logger, dotenv, initGDALOGRErrors, TempRaster, VectorBase
-from rscommons import GeopackageLayer
-from rscommons.vector_ops import difference, copy_feature_class, dissolve_feature_class, intersect_feature_classes, remove_holes_feature_class, geom_validity_fix
-from rscommons.thiessen.vor import NARVoronoi
-from rscommons.thiessen.shapes import centerline_points
-from rscommons.vbet_network import vbet_network, create_stream_size_zones, copy_vaa_attributes, join_attributes
-from rscommons.classes.raster import get_data_polygon
-
-from vbet.vbet_database import load_configuration, build_vbet_database
-from vbet.vbet_metrics import build_vbet_metric_tables
+from vbet.vbet_database import build_vbet_database, load_configuration
+from vbet.vbet_raster_ops import rasterize_attribute, raster2array, array2raster, new_raster, rasterize
+from vbet.vbet_outputs import vbet_merge
 from vbet.vbet_report import VBETReport
-from vbet.vbet_raster_ops import rasterize, raster_clean, rasterize_attribute
-from vbet.vbet_outputs import threshold, sanitize
-from vbet.vbet_centerline import vbet_centerline
+from scripts.cost_path import least_cost_path
+from scripts.raster2line import raster2line_geom
 from vbet.__version__ import __version__
 
 Path = str
+
+NCORES = "8"
 
 initGDALOGRErrors()
 
@@ -52,7 +48,6 @@ cfg = ModelConfig('http://xml.riverscapes.xyz/Projects/XSD/V1/VBET.xsd', __versi
 LayerTypes = {
     'DEM': RSLayer('DEM', 'DEM', 'Raster', 'inputs/dem.tif'),
     'SLOPE_RASTER': RSLayer('Slope Raster', 'SLOPE_RASTER', 'Raster', 'inputs/slope.tif'),
-    'HAND_RASTER': RSLayer('Hand Raster', 'HAND_RASTER', 'Raster', 'inputs/HAND.tif'),
     'TWI_RASTER': RSLayer('Topographic Wetness Index (TWI) Raster', 'TWI_RASTER', 'Raster', 'inputs/twi.tif'),
     'HILLSHADE': RSLayer('DEM Hillshade', 'HILLSHADE', 'Raster', 'inputs/dem_hillshade.tif'),
     'INPUTS': RSLayer('Inputs', 'INPUTS', 'Geopackage', 'inputs/vbet_inputs.gpkg', {
@@ -62,73 +57,45 @@ LayerTypes = {
         'CHANNEL_AREA_POLYGONS': RSLayer('Channel Area Polygons', 'CHANNEL_AREA_POLYGONS', 'Vector', 'channel_area_polygons'),
         'CATCHMENTS': RSLayer('NHD Catchments', 'CATCHMENTS', 'Vector', 'catchments'),
     }),
-    'CHANNEL_AREA_RASTER': RSLayer('Channel Area Raster', 'CHANNEL_AREA_RASTER', 'Raster', 'intermediates/channelarea.tif'),
-    # 'CHANNEL_DISTANCE': RSLayer('Channel Euclidean Distance', 'CHANNEL_DISTANCE', "Raster", "intermediates/ChannelEuclideanDist.tif"),
+    # 'CHANNEL_AREA_RASTER': RSLayer('Channel Area Raster', 'CHANNEL_AREA_RASTER', 'Raster', 'intermediates/channelarea.tif'),
     # DYNAMIC: 'DA_ZONE_<RASTER>': RSLayer('Drainage Area Zone Raster', 'DA_ZONE_RASTER', "Raster", "intermediates/.tif"),
-    'NORMALIZED_SLOPE': RSLayer('Normalized Slope', 'NORMALIZED_SLOPE', "Raster", "intermediates/nLoE_Slope.tif"),
-    'NORMALIZED_HAND': RSLayer('Normalized HAND', 'NORMALIZED_HAND', "Raster", "intermediates/nLoE_Hand.tif"),
-    # 'NORMALIZED_CHANNEL_DISTANCE': RSLayer('Normalized Channel Distance', 'NORMALIZED_CHANNEL_DISTANCE', "Raster", "intermediates/nLoE_ChannelDist.tif"),
-    'NORMALIZED_TWI': RSLayer('Normalized Topographic Wetness Index (TWI)', 'NORMALIZED_TWI', "Raster", "intermediates/nLoE_TWI.tif"),
-    'EVIDENCE_TOPO': RSLayer('Topo Evidence', 'EVIDENCE_TOPO', 'Raster', 'intermediates/Topographic_Evidence.tif'),
-    'EVIDENCE_CHANNEL': RSLayer('Channel Evidence', 'EVIDENCE_CHANNEL', 'Raster', 'intermediates/Channel_Evidence.tif'),
+    # 'NORMALIZED_SLOPE': RSLayer('Normalized Slope', 'NORMALIZED_SLOPE', "Raster", "intermediates/nLoE_Slope.tif"),
+    # 'NORMALIZED_HAND': RSLayer('Normalized HAND', 'NORMALIZED_HAND', "Raster", "intermediates/nLoE_Hand.tif"),
+    # 'NORMALIZED_TWI': RSLayer('Normalized Topographic Wetness Index (TWI)', 'NORMALIZED_TWI', "Raster", "intermediates/nLoE_TWI.tif"),
+    # 'EVIDENCE_TOPO': RSLayer('Topo Evidence', 'EVIDENCE_TOPO', 'Raster', 'intermediates/Topographic_Evidence.tif'),
+    # 'EVIDENCE_CHANNEL': RSLayer('Channel Evidence', 'EVIDENCE_CHANNEL', 'Raster', 'intermediates/Channel_Evidence.tif'),
     'INTERMEDIATES': RSLayer('Intermediates', 'Intermediates', 'Geopackage', 'intermediates/vbet_intermediates.gpkg', {
         'VBET_NETWORK': RSLayer('VBET Network', 'VBET_NETWORK', 'Vector', 'vbet_network'),
         'TRANSFORM_ZONES': RSLayer('Transform Zones', 'TRANSFORM_ZONES', 'Vector', 'transform_zones'),
-        'THIESSEN_POINTS': RSLayer('Thiessen Reach Points', 'THIESSEN_POINTS', 'Vector', 'ThiessenPoints'),
-        'THIESSEN_AREAS': RSLayer('Thiessen Reach Areas', 'THIESSEN_AREAS', 'Vector', 'ThiessenPolygonsDissolved'),
+        # 'THIESSEN_POINTS': RSLayer('Thiessen Reach Points', 'THIESSEN_POINTS', 'Vector', 'ThiessenPoints'),
+        # 'THIESSEN_AREAS': RSLayer('Thiessen Reach Areas', 'THIESSEN_AREAS', 'Vector', 'ThiessenPolygonsDissolved'),
         # 'CHANNEL_AREA_INTERSECTION': RSLayer('Channel Area Intetrsected by VBET', 'CHANNEL_AREA_INTERSECTION', 'Vector', 'channel_area_intersection')
         # We also add all tht raw thresholded shapes here but they get added dynamically later
     }),
     # Same here. Sub layers are added dynamically later.
-    'VBET_EVIDENCE': RSLayer('VBET Evidence Raster', 'VBET_EVIDENCE', 'Raster', 'outputs/VBET_Evidence.tif'),
+    'COMPOSITE_VBET_EVIDENCE': RSLayer('VBET Evidence Raster', 'VBET_EVIDENCE', 'Raster', 'outputs/VBET_Evidence.tif'),
+    'COMPOSITE_HAND_RASTER': RSLayer('Hand Raster', 'HAND_RASTER', 'Raster', 'inputs/HAND.tif'),
     'VBET_OUTPUTS': RSLayer('VBET', 'VBET_OUTPUTS', 'Geopackage', 'outputs/vbet.gpkg', {
         'VBET_FULL': RSLayer('VBET Full Extent', 'VBET_FULL', 'Vector', 'vbet_full'),
         'VBET_IA': RSLayer('VBET Inactive/Active Boundary', 'VBET_IA', 'Vector', 'vbet_ia'),
         'VBET_CHANNEL_AREA': RSLayer('VBET Channel Area', 'VBET_CHANNEL_AREA', 'Vector', 'vbet_channel_area'),
         'ACTIVE_FLOODPLAIN': RSLayer('Active Floodplain', 'ACTIVE_FLOODPLAIN', 'Vector', 'active_floodplain'),
-        'INACTIVE_FLOODPLAIN': RSLayer('Inactive Floodplain', 'INACTIVE_FLOODPLAIN', 'Vector', 'inactive_floodplain')
+        'INACTIVE_FLOODPLAIN': RSLayer('Inactive Floodplain', 'INACTIVE_FLOODPLAIN', 'Vector', 'inactive_floodplain'),
+        'VBET_CENTERLINE': RSLayer('VBET Centerline', 'VBET_CENTERLINE', 'Vector', 'vbet_centerline')
     }),
     'REPORT': RSLayer('RSContext Report', 'REPORT', 'HTMLFile', 'outputs/vbet.html')
 }
 
-flowline_fieldmap = {
-    'NHD': {
-        'ReachID': 'NHDPlusID',
-        'PathID': 'LevelPathI',
-        'DownPathID': 'DnLevelPat',
-        'Divergence': 'Divergence',
-        'StreamOrder': 'StreamOrde'},
-    'NetMap': {
-        'ReachID': 'ID',
-        'PathID': 'CHAN_ID',
-        'DownPathID': 'DOWN_ID',
-        'Divergence': '',
-        'StreamOrder': 'STRM_ORDER'},
-    'Custom': {
-        'ReachID': 'ReachID',
-        'PathID': 'PathID',
-        'DownPathID': 'DownPathID',
-        'Divergence': 'Divergence',
-        'StreamOrder': 'StreamOrder'
-    }}
 
+def vbet_centerlines(in_line_network, dem, slope, in_catchments, in_channel_area, vaa_table, project_folder, scenario_code, huc, level_paths=None, pitfill_dem=None, dinfflowdir_ang=None, dinfflowdir_slp=None, twi_raster=None, meta=None, debug=True):
 
-def vbet(huc: int, scenario_code: str, inputs: Dict[str, str], vaa_table: Path, project_folder: Path, reach_codes: List[str], meta: Dict[str, str], flowline_type: str = 'NHD', epsg=cfg.OUTPUT_EPSG, thresh_vals={'VBET_IA': 0.90, 'VBET_FULL': 0.68}):
-    """generate vbet evidence raster and threshold polygons for a watershed
-
-    Args:
-        huc (int): HUC code for watershed
-        scenario_code (str): database machine code for scenario to run
-        inputs (dict): input names and path
-        vaa_table (Path): NHD VAA table to join with flowlines
-        project_folder (Path): path for project results
-        reach_codes (List[int]): NHD reach codes for features to include in outputs
-        meta (Dict[str,str]): dictionary of riverscapes metadata key: value pairs
-    """
+    thresh_vals = {'VBET_IA': 0.90, 'VBET_FULL': 0.68}
 
     vbet_timer = time.time()
     log = Logger('VBET')
     log.info('Starting VBET v.{}'.format(cfg.version))
+
+    flowline_type = 'NHD'
 
     project, _realization, proj_nodes = create_project(huc, project_folder, [
         RSMeta('HUC{}'.format(len(huc)), str(huc)),
@@ -141,372 +108,252 @@ def vbet(huc: int, scenario_code: str, inputs: Dict[str, str], vaa_table: Path, 
         RSMeta("VBET_Inactive_Floodplain_Threshold", f"{int(thresh_vals['VBET_FULL'] * 100)}", RSMetaTypes.INT)
     ], meta)
 
-    flowline_fields = flowline_fieldmap[flowline_type]
+    inputs_gpkg = os.path.join(project_folder, LayerTypes['INPUTS'].rel_path)
+    intermediates_gpkg = os.path.join(project_folder, LayerTypes['INTERMEDIATES'].rel_path)
+    vbet_gpkg = os.path.join(project_folder, LayerTypes['VBET_OUTPUTS'].rel_path)
+    GeopackageLayer.delete(inputs_gpkg)
+    GeopackageLayer.delete(intermediates_gpkg)
+    GeopackageLayer.delete(vbet_gpkg)
 
-    vbet_summary_field = 'vbet_level_path'  # This value is in vbet_metrics.sql, modify with care...
+    line_network_features = os.path.join(inputs_gpkg, LayerTypes['INPUTS'].sub_layers['FLOWLINES'].rel_path)
+    copy_feature_class(in_line_network, line_network_features)
+    catchment_features = os.path.join(inputs_gpkg, LayerTypes['INPUTS'].sub_layers['CATCHMENTS'].rel_path)
+    copy_feature_class(in_catchments, catchment_features)
+    channel_area = os.path.join(inputs_gpkg, LayerTypes['INPUTS'].sub_layers['CHANNEL_AREA_POLYGONS'].rel_path)
+    copy_feature_class(in_channel_area, channel_area)
 
-    # Input Preparation
-    # Make sure we're starting with a fresh slate of new geopackages
-    inputs_gpkg_path = os.path.join(project_folder, LayerTypes['INPUTS'].rel_path)
-    intermediates_gpkg_path = os.path.join(project_folder, LayerTypes['INTERMEDIATES'].rel_path)
-    vbet_path = os.path.join(project_folder, LayerTypes['VBET_OUTPUTS'].rel_path)
-    GeopackageLayer.delete(inputs_gpkg_path)
-    GeopackageLayer.delete(intermediates_gpkg_path)
-    GeopackageLayer.delete(vbet_path)
+    build_vbet_database(inputs_gpkg)
+    vbet_run = load_configuration(scenario_code, inputs_gpkg)
 
-    raster_extent = MultiPolygon(get_data_polygon(inputs['DEM']))
-    raster_extent_geom = VectorBase.shapely2ogr(raster_extent)
+    vaa_table_name = copy_vaa_attributes(line_network_features, vaa_table)
+    line_network = join_attributes(inputs_gpkg, "flowlines_vaa", os.path.basename(line_network_features), vaa_table_name, 'NHDPlusID', ['LevelPathI', 'DnLevelPat', 'UpLevelPat', 'Divergence'], 4326)
+    catchments = join_attributes(inputs_gpkg, "catchments_vaa", os.path.basename(catchment_features), vaa_table_name, 'NHDPlusID', ['LevelPathI', 'DnLevelPat', 'UpLevelPat', 'Divergence'], 4326, geom_type='POLYGON')
 
-    project_inputs = {}
-    for input_name, input_path in inputs.items():
-        if os.path.splitext(input_path)[1] in ['.tif', '.tiff', '.TIF', '.TIFF']:
-            _proj_slope_node, project_inputs[input_name] = project.add_project_raster(proj_nodes['Inputs'], LayerTypes[input_name], input_path)
-        else:
-            project_path = os.path.join(inputs_gpkg_path, LayerTypes['INPUTS'].sub_layers[input_name].rel_path)
-            copy_feature_class(input_path, project_path, epsg=epsg, clip_shape=raster_extent_geom)
-            project_inputs[input_name] = project_path
-    project.add_project_geopackage(proj_nodes['Inputs'], LayerTypes['INPUTS'])
+    get_channel_level_path(channel_area, line_network, vaa_table)
 
-    vaa_table_name = copy_vaa_attributes(project_inputs['FLOWLINES'], vaa_table)
-    if flowline_type == 'NHD':
-        vaa_fields = ['LevelPathI', 'Divergence', 'DnLevelPat']  # + ["StreamOrde"]  # + ['HydroSeq', 'DnHydroSeq', 'UpHydroSeq']
-        flowlines_vaa_path = join_attributes(inputs_gpkg_path, "Flowlines_VAA", os.path.basename(project_inputs['FLOWLINES']), vaa_table_name, 'NHDPlusID', vaa_fields, epsg)
-        catchments_vaa_path = join_attributes(inputs_gpkg_path, "Catchments_VAA", os.path.basename(project_inputs['CATCHMENTS']), vaa_table_name, 'NHDPlusID', vaa_fields, epsg, geom_type='POLYGON')
-    else:
-        flowlines_vaa_path = project_inputs['FLOWLINES']
+    catchments_path = os.path.join(intermediates_gpkg, 'transform_zones')
+    vaa_table_path = os.path.join(inputs_gpkg, vaa_table_name)
+    create_stream_size_zones(catchments, vaa_table_path, 'NHDPlusID', 'StreamOrde', vbet_run['Zones'], catchments_path)
 
-    # Build Transformation Tables
-    build_vbet_database(inputs_gpkg_path)
-    # with GeopackageLayer(project_inputs['FLOWLINES']) as lyr:
-    #     degree_factor = lyr.rough_convert_metres_to_vector_units(1)
-
-    # Load configuration from table
-    vbet_run = load_configuration(scenario_code, inputs_gpkg_path)
-
-    # Create a copy of the flow lines with just the perennial and also connectors inside flow areas
-    log.info('Building vbet network')
-    network_path = os.path.join(intermediates_gpkg_path, LayerTypes['INTERMEDIATES'].sub_layers['VBET_NETWORK'].rel_path)
-    vbet_network(flowlines_vaa_path, project_inputs['FLOW_AREAS'], network_path, epsg, reach_codes)
-
-    # Create Zones
-    log.info('Building drainage area zones')
-    catchments_path = os.path.join(intermediates_gpkg_path, LayerTypes['INTERMEDIATES'].sub_layers['TRANSFORM_ZONES'].rel_path)
-    vaa_table_path = os.path.join(inputs_gpkg_path, vaa_table_name)
-    create_stream_size_zones(project_inputs['CATCHMENTS'], vaa_table_path, 'NHDPlusID', 'StreamOrde', vbet_run['Zones'], catchments_path)  # TODO not fully generic here. Relies on NHD catchments and vaa attributes
-
-    # Create Scenario Input Rasters
     in_rasters = {}
-    out_rasters = {}
-    if 'Slope' in vbet_run['Inputs']:
-        log.info("Adding Slope Input")
-        in_rasters['Slope'] = project_inputs['SLOPE_RASTER']
-        out_rasters['NORMALIZED_SLOPE'] = os.path.join(project_folder, LayerTypes['NORMALIZED_SLOPE'].rel_path)
+    in_rasters['Slope'] = slope
 
-    # Rasterize the channel polygon and write to raster
-    log.info('Writing channel raster using slope as a template')
-    channel_area_raster = os.path.join(project_folder, LayerTypes['CHANNEL_AREA_RASTER'].rel_path)
-    rasterize(project_inputs['CHANNEL_AREA_POLYGONS'], channel_area_raster, project_inputs['SLOPE_RASTER'], all_touched=True)
-    project.add_project_raster(proj_nodes['Intermediates'], LayerTypes['CHANNEL_AREA_RASTER'])
-    in_rasters['Channel'] = channel_area_raster
+    # generate top level taudem products if they do not exist
+    if pitfill_dem is None:
+        pitfill_dem = os.path.join(project_folder, 'pitfill.tif')
+        pitfill_status = run_subprocess(project_folder, ["mpiexec", "-n", NCORES, "pitremove", "-z", dem, "-fel", pitfill_dem])
+        if pitfill_status != 0 or not os.path.isfile(pitfill_dem):
+            raise Exception('TauDEM: pitfill failed')
 
-    # Generate HAND from dem and rasterized flow polygons
-    if 'HAND' in vbet_run['Inputs']:
-        hand_raster = os.path.join(project_folder, LayerTypes['HAND_RASTER'].rel_path)
-        twi_raster = os.path.join(project_folder, LayerTypes['TWI_RASTER'].rel_path)
-        in_rasters['HAND'] = hand_raster
-        out_rasters['NORMALIZED_HAND'] = os.path.join(project_folder, LayerTypes['NORMALIZED_HAND'].rel_path)
+    if not all([dinfflowdir_ang, dinfflowdir_slp]):
+        dinfflowdir_slp = os.path.join(project_folder, 'dinfflowdir_slp.tif')
+        dinfflowdir_ang = os.path.join(project_folder, 'dinfflowdir_ang.tif')
+        dinfflowdir_status = run_subprocess(project_folder, ["mpiexec", "-n", NCORES, "dinfflowdir", "-fel", pitfill_dem, "-ang", dinfflowdir_ang, "-slp", dinfflowdir_slp])
+        if dinfflowdir_status != 0 or not os.path.isfile(dinfflowdir_ang):
+            raise Exception('TauDEM: dinfflowdir failed')
 
-        in_rasters['TWI'] = twi_raster
-        out_rasters['NORMALIZED_TWI'] = os.path.join(project_folder, LayerTypes['NORMALIZED_TWI'].rel_path)
+    if not twi_raster:
+        twi_raster = os.path.join(project_folder, 'local_twi')
+        sca = os.path.join(project_folder, 'sca.tif')
+        twi_status = run_subprocess(project_folder, ["mpiexec", "-n", NCORES, "twi", "-slp", dinfflowdir_slp, "-sca", sca, '-twi', twi_raster])
+        if twi_status != 0 or not os.path.isfile(twi_raster):
+            raise Exception('TauDEM: TWI failed')
+    in_rasters['TWI'] = twi_raster
 
-    # Generate da Zone rasters
     for zone in vbet_run['Zones']:
         log.info(f'Rasterizing stream transform zones for {zone}')
         raster_name = os.path.join(project_folder, 'intermediates', f'{zone.lower()}_transform_zones.tif')
-        rasterize_attribute(catchments_path, raster_name, project_inputs['SLOPE_RASTER'], f'{zone}_Zone')
+        rasterize_attribute(catchments_path, raster_name, dem, f'{zone}_Zone')
         in_rasters[f'TRANSFORM_ZONE_{zone}'] = raster_name
         transform_zone_rs = RSLayer(f'Transform Zones for {zone}', f'TRANSFORM_ZONE_{zone.upper()}', 'Raster', raster_name)
         project.add_project_raster(proj_nodes['Intermediates'], transform_zone_rs)
 
-    for raster_name in ['EVIDENCE_TOPO', 'EVIDENCE_CHANNEL']:
-        out_rasters[raster_name] = os.path.join(project_folder, LayerTypes[raster_name].rel_path)
-    evidence_raster = os.path.join(project_folder, LayerTypes['VBET_EVIDENCE'].rel_path)
+    # run for orphan waterbodies??
 
-    # Open evidence rasters concurrently. We're looping over windows so this shouldn't affect
-    # memory consumption too much
-    read_rasters = {name: rasterio.open(raster) for name, raster in in_rasters.items()}
-    out_meta = read_rasters['Slope'].meta
-    out_meta['driver'] = 'GTiff'
-    out_meta['count'] = 1
-    out_meta['compress'] = 'deflate'
+    # Initialize Outputs
+    output_centerlines = os.path.join(vbet_gpkg, "vbet_centerlines")
+    output_vbet = os.path.join(vbet_gpkg, LayerTypes["OUTPUTS"].sub_layers['VBET_FULL'])
+    output_vbet_ia = os.path.join(vbet_gpkg, LayerTypes['OUTPUTS'].sub_layers['VBET_IA'])
+    with GeopackageLayer(output_centerlines, write=True) as lyr_cl_init, \
+        GeopackageLayer(output_vbet, write=True) as lyr_vbet_init, \
+        GeopackageLayer(output_vbet_ia, write=True) as lyr_active_vbet_init, \
+            GeopackageLayer(line_network) as lyr_ref:
+        fields = {'LevelPathI': ogr.OFTString}
+        lyr_cl_init.create_layer(ogr.wkbMultiLineString, spatial_ref=lyr_ref.spatial_ref, fields=fields)
+        lyr_vbet_init.create_layer(ogr.wkbPolygon, spatial_ref=lyr_ref.spatial_ref, fields=fields)
+        lyr_active_vbet_init.create_layer(ogr.wkbPolygon, spatial_ref=lyr_ref.spatial_ref, fields=fields)
 
-    # We use this to buffer the output
-    cell_size = abs(read_rasters['Slope'].get_transform()[1])
+    out_hand = os.path.join(project_folder, 'outputs', "composite_hand.tif")
+    out_vbet_evidence = os.path.join(project_folder, 'outputs', 'composite_vbet_evidence.tif')
 
-    write_rasters = {name: rasterio.open(raster, 'w', **out_meta) for name, raster in out_rasters.items()}
-    write_rasters['VBET_EVIDENCE'] = rasterio.open(evidence_raster, 'w', **out_meta)
+    # Generate the list of level paths to run, sorted by ascending order and optional user filter
+    level_paths_to_run = []
+    with GeopackageLayer(line_network) as line_lyr:
+        for feat, *_ in line_lyr.iterate_features():
+            level_path = feat.GetField('LevelPathI')
+            level_paths_to_run.append(str(int(level_path)))
+    level_paths_to_run = list(set(level_paths_to_run))
+    if level_paths:
+        level_paths_to_run = [level_path for level_path in level_paths_to_run if level_path in level_paths]
+    level_paths_to_run.sort(reverse=False)
+    level_paths_to_run.append(None)
 
-    progbar = ProgressBar(len(list(read_rasters['Slope'].block_windows(1))), 50, "Calculating evidence layer")
-    counter = 0
-    # Again, these rasters should be orthogonal so their windows should also line up
-    for _ji, window in read_rasters['Slope'].block_windows(1):
-        progbar.update(counter)
-        counter += 1
-        block = {block_name: raster.read(1, window=window, masked=True) for block_name, raster in read_rasters.items()}
+    project.add_project_geopackage(proj_nodes['Inputs'], LayerTypes['INPUTS'])
 
-        normalized = {}
-        for name in vbet_run['Inputs']:
-            if name in vbet_run['Zones']:
-                transforms = [np.ma.MaskedArray(transform(block[name].data), mask=block[name].mask) for transform in vbet_run['Transforms'][name]]
-                normalized[name] = np.ma.MaskedArray(np.choose(block[f'TRANSFORM_ZONE_{name}'].data, transforms, mode='clip'), mask=block[name].mask)
+    # iterate for each level path
+    for level_path in level_paths_to_run:
+
+        temp_folder = os.path.join(project_folder, 'temp', f'levelpath_{level_path}')
+        safe_makedirs(temp_folder)
+
+        sql = f"LevelPathI = {level_path}" if level_path is not None else "LevelPathI is NULL"
+        # Select channel areas that intersect flow lines
+        level_path_polygons = os.path.join(project_folder, 'temp', 'channel_polygons.gpkg', f'level_path_{level_path}')
+        copy_feature_class(channel_area, level_path_polygons, attribute_filter=sql)
+
+        # current_vbet = collect_feature_class(output_vbet)
+        # if current_vbet is not None:
+        #     channel_polygons = collect_feature_class(level_path_polygons)
+        #     if current_vbet.Contains(channel_polygons):
+        #         continue
+
+        evidence_raster = os.path.join(temp_folder, f'vbet_evidence_{level_path}.tif')
+        rasterized_channel = os.path.join(temp_folder, f'rasterized_channel_{level_path}.tif')
+        rasterize(level_path_polygons, rasterized_channel, dem, all_touched=True)
+        in_rasters['Channel'] = rasterized_channel
+
+        # log.info("Generating HAND")
+        hand_raster = os.path.join(temp_folder, f'local_hand_{level_path}.tif')
+        dinfdistdown_status = run_subprocess(project_folder, ["mpiexec", "-n", NCORES, "dinfdistdown", "-ang", dinfflowdir_ang, "-fel", pitfill_dem, "-src", rasterized_channel, "-dd", hand_raster, "-m", "ave", "v"])
+        if dinfdistdown_status != 0 or not os.path.isfile(hand_raster):
+            raise Exception('TauDEM: dinfdistdown failed')
+        in_rasters['HAND'] = hand_raster
+
+        # Open evidence rasters concurrently. We're looping over windows so this shouldn't affect
+        # memory consumption too much
+        read_rasters = {name: rasterio.open(raster) for name, raster in in_rasters.items()}
+        out_meta = read_rasters['Slope'].meta
+        out_meta['driver'] = 'GTiff'
+        out_meta['count'] = 1
+        out_meta['compress'] = 'deflate'
+
+        write_rasters = {}  # {name: rasterio.open(raster, 'w', **out_meta) for name, raster in out_rasters.items()}
+        write_rasters['VBET_EVIDENCE'] = rasterio.open(evidence_raster, 'w', **out_meta)
+
+        progbar = ProgressBar(len(list(read_rasters['Slope'].block_windows(1))), 50, "Calculating evidence layer")
+        counter = 0
+        # Again, these rasters should be orthogonal so their windows should also line up
+        for _ji, window in read_rasters['HAND'].block_windows(1):
+            progbar.update(counter)
+            counter += 1
+            block = {block_name: raster.read(1, window=window, masked=True) for block_name, raster in read_rasters.items()}
+
+            normalized = {}
+            for name in vbet_run['Inputs']:
+                if name in vbet_run['Zones']:
+                    transforms = [np.ma.MaskedArray(transform(block[name].data), mask=block[name].mask) for transform in vbet_run['Transforms'][name]]
+                    normalized[name] = np.ma.MaskedArray(np.choose(block[f'TRANSFORM_ZONE_{name}'].data, transforms, mode='clip'), mask=block[name].mask)
+                else:
+                    normalized[name] = np.ma.MaskedArray(vbet_run['Transforms'][name][0](block[name].data), mask=block[name].mask)
+
+            fvals_topo = np.ma.mean([normalized['Slope'], normalized['HAND'], normalized['TWI']], axis=0)
+            fvals_channel = 0.995 * block['Channel']
+            fvals_evidence = np.maximum(fvals_topo, fvals_channel)
+
+            write_rasters['VBET_EVIDENCE'].write(np.ma.filled(np.float32(fvals_evidence), out_meta['nodata']), window=window, indexes=1)
+        write_rasters['VBET_EVIDENCE'].close()
+
+        # Generate VBET Polygon
+        valley_bottom_raster = os.path.join(temp_folder, f'valley_bottom_{level_path}.tif')
+        generate_vbet_polygon(evidence_raster, rasterized_channel, hand_raster, valley_bottom_raster, temp_folder)
+        vbet_polygon = os.path.join(temp_folder, f'valley_bottom_{level_path}.shp')
+        polygonize(valley_bottom_raster, 1, vbet_polygon, epsg=4326)
+        # Add to existing feature class
+        polygon = vbet_merge(vbet_polygon, output_vbet, level_path=level_path)
+
+        active_valley_bottom_raster = os.path.join(temp_folder, f'active_valley_bottom_{level_path}.tif')
+        generate_vbet_polygon(evidence_raster, rasterized_channel, hand_raster, active_valley_bottom_raster, temp_folder, threshold=0.90)
+        active_vbet_polygon = os.path.join(temp_folder, f'active_valley_bottom_{level_path}.shp')
+        polygonize(active_valley_bottom_raster, 1, active_vbet_polygon, epsg=4326)
+        # Add to existing feature class
+        active_polygon = vbet_merge(active_vbet_polygon, output_vbet_ia, level_path=level_path)
+
+        # Generate centerline for level paths only
+        if level_path is not None:
+            # Generate Centerline from Cost Path
+            cost_path_raster = os.path.join(temp_folder, f'cost_path_{level_path}.tif')
+            generate_centerline_surface(valley_bottom_raster, cost_path_raster, temp_folder)
+            centerline_raster = os.path.join(temp_folder, f'centerline_{level_path}.tif')
+            coords = get_endpoints(line_network, 'LevelPathI', level_path)
+            least_cost_path(cost_path_raster, centerline_raster, coords[0], coords[1])
+            centerline_full = raster2line_geom(centerline_raster, 1)
+
+            if polygon is not None:
+                centerline_intersected = polygon.Intersection(centerline_full)
+                if centerline_intersected.GetGeometryName() == 'GeometryCollection':
+                    for line in centerline_intersected:
+                        centerline = ogr.Geometry(ogr.wkbMultiLineString)
+                        if line.GetGeometryName() == 'LineString':
+                            centerline.AddGeometry(line)
+                else:
+                    centerline = centerline_intersected
             else:
-                normalized[name] = np.ma.MaskedArray(vbet_run['Transforms'][name][0](block[name].data), mask=block[name].mask)
+                centerline = centerline_full
 
-        fvals_topo = np.ma.mean([normalized['Slope'], normalized['HAND'], normalized['TWI']], axis=0)
-        fvals_channel = 0.995 * block['Channel']
-        fvals_evidence = np.maximum(fvals_topo, fvals_channel)
+            with GeopackageLayer(output_centerlines, write=True) as lyr_cl:
+                out_feature = ogr.Feature(lyr_cl.ogr_layer_def)
+                out_feature.SetGeometry(centerline)
+                out_feature.SetField('LevelPathI', str(level_path))
+                lyr_cl.ogr_layer.CreateFeature(out_feature)
+                out_feature = None
 
-        # Fill the masked values with the appropriate nodata vals
-        # Unthresholded in the base band (mostly for debugging)
-        write_rasters['VBET_EVIDENCE'].write(np.ma.filled(np.float32(fvals_evidence), out_meta['nodata']), window=window, indexes=1)
+        # clean up rasters
+        rio_vbet = rasterio.open(valley_bottom_raster)
 
-        write_rasters['NORMALIZED_SLOPE'].write(normalized['Slope'].astype('float32').filled(out_meta['nodata']), window=window, indexes=1)
-        write_rasters['NORMALIZED_HAND'].write(normalized['HAND'].astype('float32').filled(out_meta['nodata']), window=window, indexes=1)
-        write_rasters['NORMALIZED_TWI'].write(normalized['TWI'].astype('float32').filled(out_meta['nodata']), window=window, indexes=1)
+        for out_raster, in_raster in {out_hand: hand_raster, out_vbet_evidence: evidence_raster}.items():
+            if os.path.exists(out_raster):
+                out_temp = os.path.join(temp_folder, 'temp_raster')
+                rio_dest = rasterio.open(out_raster)
+                out_meta = rio_dest.meta
+                out_meta['driver'] = 'GTiff'
+                out_meta['count'] = 1
+                out_meta['compress'] = 'deflate'
+                rio_temp = rasterio.open(out_temp, 'w', **out_meta)
+                rio_source = rasterio.open(in_raster)
+                for _ji, window in rio_source.block_windows(1):
+                    # progbar.update(counter)
+                    # counter += 1
+                    array_vbet_mask = np.ma.MaskedArray(rio_vbet.read(1, window=window, masked=True).data)
+                    array_source = np.ma.MaskedArray(rio_source.read(1, window=window, masked=True).data, mask=array_vbet_mask.mask)
+                    array_dest = np.ma.MaskedArray(rio_dest.read(1, window=window, masked=True).data, mask=array_vbet_mask.mask)
+                    array_out = np.choose(array_vbet_mask, [array_dest, array_source])
 
-        write_rasters['EVIDENCE_CHANNEL'].write(np.ma.filled(np.float32(fvals_channel), out_meta['nodata']), window=window, indexes=1)
-        write_rasters['EVIDENCE_TOPO'].write(np.ma.filled(np.float32(fvals_topo), out_meta['nodata']), window=window, indexes=1)
-    progbar.finish()
+                    rio_temp.write(np.ma.filled(np.float32(array_out), out_meta['nodata']), window=window, indexes=1)
+                rio_temp.close()
+                rio_dest.close()
+                rio_source.close()
+                shutil.copyfiles(out_temp, out_raster)
+            else:
+                rio_source = rasterio.open(in_raster)
+                out_meta = rio_source.meta
+                out_meta['driver'] = 'GTiff'
+                out_meta['count'] = 1
+                out_meta['compress'] = 'deflate'
+                rio_dest = rasterio.open(out_raster, 'w', **out_meta)
+                for _ji, window in rio_source.block_windows(1):
+                    array_vbet_mask = np.ma.MaskedArray(rio_vbet.read(1, window=window, masked=True).data)
+                    array_source = np.ma.MaskedArray(rio_source.read(1, window=window, masked=True).data, mask=array_vbet_mask.mask)
+                    rio_dest.write(np.ma.filled(np.float32(array_source), out_meta['nodata']), window=window, indexes=1)
+                rio_dest.close()
+                rio_source.close()
 
-    # Close all rasters here
-    for raster_obj in list(read_rasters.values()) + list(write_rasters.values()):
-        raster_obj.close()
-
-    # The remaining rasters get added to the project
-    for raster_name in out_rasters:
-        project.add_project_raster(proj_nodes["Intermediates"], LayerTypes[raster_name])
-    project.add_project_raster(proj_nodes['Outputs'], LayerTypes['VBET_EVIDENCE'])
-
-    min_geom_size = 10 * (cell_size ** 2)
-
-    # log.info('Subdividing the network along regular intervals')
-    # flowline_thiessen_points_groups = centerline_points(network_path, distance=degree_factor * 10, fields=[flowline_fields['PathID']], divergence_field=flowline_fields['Divergence'], downlevel_field=flowline_fields['DownPathID'])  # 'LevelPathI'
-    # flowline_thiessen_points = [pt for group in flowline_thiessen_points_groups.values() for pt in group]
-
-    # Exterior is the shell and there is only ever 1
-    # log.info("Creating Thiessen Polygons")
-    # myVorL = NARVoronoi(flowline_thiessen_points)
-
-    # Generate Thiessen Polys
-    # myVorL.createshapes()
-
-    # with GeopackageLayer(network_path) as flow_lyr:
-    #     # Set the output spatial ref as this for the whole project
-    #     out_srs = flow_lyr.spatial_ref
-
-    # Dissolve by flowlines
-    # log.info("Dissolving Thiessen Polygons")
-    # dissolved_polys = myVorL.dissolve_by_property(flowline_fields['PathID'])  # 'LevelPathI'
-
-    # dissolved_attributes = {flowline_fields['PathID']: ogr.OFTString}  # 'LevelPathI'
-
-    # simple_save([{'geom': pt.point} for pt in flowline_thiessen_points], ogr.wkbPoint, out_srs, "ThiessenPoints", intermediates_gpkg_path)
-    # simple_save([{'geom': g, 'attributes': {flowline_fields['PathID']: k}} for k, g in dissolved_polys.items()], ogr.wkbPolygon, out_srs, "ThiessenPolygonsDissolved", intermediates_gpkg_path, dissolved_attributes)
-
-    # thiessen_fc = os.path.join(intermediates_gpkg_path, "ThiessenPolygonsDissolved")
-    # modified_catchments_vaa_path = os.path.join(intermediates_gpkg_path, 'catchments_vaa_modified_path')
-    # copy_feature_class(catchments_vaa_path, modified_catchments_vaa_path)
-
-    # with GeopackageLayer(modified_catchments_vaa_path, write=True) as lyr:
-    #     for feat, *_ in lyr.iterate_features("Updating divergent reach paths..."):
-    #         if feat.GetField("Divergence") == 2:
-    #             new_feat = feat.Clone()
-    #             output = feat.GetField("DnLevelPat")
-    #             new_feat.SetField("LevelPathI", output)
-    #             lyr.ogr_layer.SetFeature(new_feat)
-
-    catchments_dissolved_path = os.path.join(intermediates_gpkg_path, "catchments_dissolved")
-    dissolve_feature_class(catchments_vaa_path, catchments_dissolved_path, epsg, vbet_summary_field)
-
-    vbet_threshold = {}
-    for str_val, thr_val in thresh_vals.items():
-
-        plgnize_id = f'THRESH_{int(thr_val * 100)}'
-        with TempRaster(f'vbet_raw_thresh_{int(thr_val * 100)}') as tmp_raw_thresh, \
-                TempRaster(f'vbet_cleaned_thresh_{int(thr_val * 100)}') as tmp_cleaned_thresh:
-
-            log.debug('Temporary threshold raster: {}'.format(tmp_raw_thresh.filepath))
-            threshold(evidence_raster, thr_val, tmp_raw_thresh.filepath)
-            raster_clean(tmp_raw_thresh.filepath, tmp_cleaned_thresh.filepath, buffer_pixels=1)
-
-            # Threshold and VBET output layers
-            plgnize_lyr = RSLayer(f'Raw Threshold at {int(thr_val * 100)}%', plgnize_id, 'Vector', plgnize_id.lower())
-            LayerTypes['INTERMEDIATES'].add_sub_layer(plgnize_id, plgnize_lyr)
-            vbet_lyr = LayerTypes['VBET_OUTPUTS'].sub_layers[str_val]
-
-            with rasterio.open(tmp_cleaned_thresh.filepath, 'r') as raster:
-                with GeopackageLayer(catchments_dissolved_path, write=True) as lyr_reaches, \
-                        GeopackageLayer(project_inputs['CHANNEL_AREA_POLYGONS']) as lyr_channel_area_polygons, \
-                        GeopackageLayer(os.path.join(intermediates_gpkg_path, plgnize_lyr.rel_path), write=True) as lyr_output:
-
-                    lyr_output.create_layer_from_ref(lyr_reaches)
-
-                    out_layer_defn = lyr_output.ogr_layer.GetLayerDefn()
-                    field_count = out_layer_defn.GetFieldCount()
-
-                    rejected_geoms = []
-
-                    lyr_output.ogr_layer.StartTransaction()
-
-                    for reach_feat, *_ in lyr_reaches.iterate_features("Processing Reaches"):
-                        reach_attributes = {}
-                        for n in range(field_count):
-                            field = out_layer_defn.GetFieldDefn(n)
-                            value = reach_feat.GetField(field.name)
-                            reach_attributes[field.name] = value
-
-                        geom = reach_feat.GetGeometryRef()
-                        buff = geom.Buffer(cell_size)
-                        geom_json = buff.ExportToJson()
-                        poly = json.loads(geom_json)
-                        data, mask_transform = rasterio.mask.mask(raster, [poly], crop=True)
-
-                        if all(x > 0 for x in data.shape):
-                            out_shapes = list(g for g, v in shapes(data, transform=mask_transform) if v == 1)
-                            for out_shape in out_shapes:
-                                shape_intersected = False
-                                out_geom = ogr.CreateGeometryFromJson(json.dumps(out_shape))
-                                if not out_geom.IsValid():
-                                    out_geom = geom_validity_fix(out_geom)
-                                out_feat = ogr.Feature(out_layer_defn)
-                                out_feat.SetGeometry(out_geom)
-                                for field, value in reach_attributes.items():
-                                    out_feat.SetField(field, value)
-                                for test_feat, *_ in lyr_channel_area_polygons.iterate_features(clip_shape=out_geom):
-                                    test_geom = test_feat.GetGeometryRef()
-                                    if test_geom.Intersects(out_geom):
-                                        shape_intersected = True
-                                        lyr_output.ogr_layer.CreateFeature(out_feat)
-                                        break
-                                if shape_intersected is False:
-                                    rejected_geoms.append(out_feat)
-                    lyr_output.ogr_layer.CommitTransaction()
-
-                    save_feats = []
-                    iterate = True
-                    while iterate is True:
-                        log.info(f"Checking {len(rejected_geoms)} features for inclusion by adjacency")
-                        iterate = False
-                        rejected_geoms2 = []
-                        for out_feat in rejected_geoms:
-                            shape_intersected = False
-                            out_geom = out_feat.GetGeometryRef()
-                            for test_feat, *_ in lyr_output.iterate_features(clip_shape=out_geom):
-                                test_geom = test_feat.GetGeometryRef()
-                                if test_geom.Intersects(out_geom):  # .Buffer(0.0001)
-                                    save_feats.append(out_feat)
-                                    iterate = True
-                                    shape_intersected = True
-                                    break
-                            if shape_intersected is False:
-                                rejected_geoms2.append(out_feat)
-                        rejected_geoms = rejected_geoms2
-
-                    lyr_output.ogr_layer.StartTransaction()
-                    for feat in save_feats:
-                        lyr_output.ogr_layer.CreateFeature(feat)
-                    lyr_output.ogr_layer.CommitTransaction()
-
-        # Now the final sanitization
-        # sanitize(
-        #     str_val,
-        #     '{}/{}'.format(intermediates_gpkg_path, plgnize_lyr.rel_path),
-        #     '{}/{}'.format(vbet_path, vbet_lyr.rel_path),
-        #     buff_dist,
-        #     network_path
-        # )
-        # log.info('Completed thresholding at {}'.format(thr_val))
-
-        remove_holes_feature_class(os.path.join(intermediates_gpkg_path, plgnize_lyr.rel_path), os.path.join(vbet_path, vbet_lyr.rel_path), min_geom_size, min_geom_size)
-        vbet_threshold[str_val] = os.path.join(vbet_path, vbet_lyr.rel_path)
-
-    # Geomorphic Layers
-    vbet_channel_area = os.path.join(vbet_path, LayerTypes['VBET_OUTPUTS'].sub_layers['VBET_CHANNEL_AREA'].rel_path)
-    with GeopackageLayer(vbet_channel_area, write=True) as output_lyr, \
-        GeopackageLayer(vbet_threshold['VBET_FULL']) as clipping_lyr, \
-            GeopackageLayer(project_inputs['CHANNEL_AREA_POLYGONS']) as target_layer:  # use catchments? likely has slightly different edges
-
-        output_lyr.create_layer(target_layer.ogr_geom_type, spatial_ref=target_layer.spatial_ref)
-        clipping_lyr_defn = clipping_lyr.ogr_layer_def
-
-        for i in range(clipping_lyr_defn.GetFieldCount()):
-            field_defn = clipping_lyr_defn.GetFieldDefn(i)
-            if field_defn.GetName() == vbet_summary_field:
-                output_lyr.ogr_layer.CreateField(field_defn)
-        output_lyr.ogr_layer.StartTransaction()
-        for clipping_feat, *_ in clipping_lyr.iterate_features('Finding features'):
-            clipping_geom = clipping_feat.GetGeometryRef().Clone()
-            clip_value = clipping_feat.GetField(vbet_summary_field)
-
-            for target_feat, *_ in target_layer.iterate_features(clip_shape=clipping_geom):
-                target_geom = target_feat.GetGeometryRef().Clone()
-                if not target_geom.IsValid():
-                    target_geom = geom_validity_fix(target_geom)
-                out_geom = clipping_geom.Intersection(target_geom)
-                if out_geom.GetGeometryType() in VectorBase.POLY_TYPES:
-                    if not out_geom.IsValid():
-                        out_geom = geom_validity_fix(out_geom)
-                    out_feat = ogr.Feature(output_lyr.ogr_layer_def)
-                    out_feat.SetGeometry(out_geom)
-                    out_feat.SetField(vbet_summary_field, clip_value)
-                    output_lyr.ogr_layer.CreateFeature(out_feat)
-        output_lyr.ogr_layer.CommitTransaction()
-
-    active_floodplain = os.path.join(vbet_path, LayerTypes['VBET_OUTPUTS'].sub_layers['ACTIVE_FLOODPLAIN'].rel_path)
-    difference(vbet_channel_area, vbet_threshold['VBET_IA'], active_floodplain, epsg)
-
-    inactive_floodplain = os.path.join(vbet_path, LayerTypes['VBET_OUTPUTS'].sub_layers['INACTIVE_FLOODPLAIN'].rel_path)
-    difference(vbet_threshold['VBET_IA'], vbet_threshold['VBET_FULL'], inactive_floodplain, epsg)
-
-    # Area Calculations
-    for layer in [active_floodplain, vbet_channel_area, inactive_floodplain, vbet_threshold['VBET_IA'], vbet_threshold['VBET_FULL']]:
-        log.info(f'Calcuating area for {layer}')
-        calculate_area(layer, "area_ha")
-
-    # TODO is this redundant now that we clip channel area to vbet? can we join those features like we join the floodplain layers?
-    with GeopackageLayer(vbet_threshold['VBET_FULL'], write=True) as lyr_vbet, \
-            GeopackageLayer(vbet_channel_area) as lyr_channel:
-
-        if lyr_vbet.ogr_layer.GetLayerDefn().GetFieldIndex('active_channel_ha') < 0:
-            lyr_vbet.create_field('active_channel_ha', ogr.OFTReal)
-        srs = lyr_vbet.get_srs_from_epsg(cfg.OUTPUT_EPSG)
-        _sr, transform = lyr_vbet.get_transform_from_epsg(srs, 5070)
-        lyr_vbet.ogr_layer.StartTransaction()
-        for feat, *_ in lyr_vbet.iterate_features("Intersecting vbet features with channel area"):
-            vbet_geom = feat.GetGeometryRef().Clone()
-            sum_area = 0.0
-            for channel_feat, *_ in lyr_channel.iterate_features(clip_shape=vbet_geom):
-                channel_geom = channel_feat.GetGeometryRef().Clone()
-                clip_geom = vbet_geom.Intersection(channel_geom)
-                clip_geom.Transform(transform)
-                area = clip_geom.GetArea()
-                hectares = area / 10000.0
-                sum_area = sum_area + hectares
-            feat.SetField('active_channel_ha', hectares)
-            lyr_vbet.ogr_layer.SetFeature(feat)
-        lyr_vbet.ogr_layer.CommitTransaction()
-    # End of redundant section...
-
-    build_vbet_metric_tables(vbet_path)
-
-    # Generate Centerline
-    # centerline_lyr = RSLayer('VBET Centerlines', 'VBET_CENTERLINES', 'Vector', 'vbet_centerlines')
-    # log.info('Creating a centerlines')
-    # LayerTypes['VBET_OUTPUTS'].add_sub_layer('VBET_CENTERLINES', centerline_lyr)
-    # centerline = os.path.join(vbet_path, centerline_lyr.rel_path)
-    # vbet_centerline(network_path, os.path.join(vbet_path, 'vbet_68'), centerline)
+        if debug is False:
+            os.rmdir(temp_folder)
 
     # Now add our Geopackages to the project XML
+    project.add_project_raster(proj_nodes['Outputs'], LayerTypes['COMPOSITE_VBET_EVIDENCE'])
+    project.add_project_raster(proj_nodes['Outputs'], LayerTypes['COMPOSITE_HAND_RASTER'])
+
     project.add_project_geopackage(proj_nodes['Intermediates'], LayerTypes['INTERMEDIATES'])
     project.add_project_geopackage(proj_nodes['Outputs'], LayerTypes['VBET_OUTPUTS'])
 
@@ -526,60 +373,68 @@ def vbet(huc: int, scenario_code: str, inputs: Dict[str, str], vaa_table: Path, 
     log.info('VBET Completed Successfully')
 
 
-def simple_save(list_geoms, ogr_type, srs, layer_name, gpkg_path, attributes={}):
+def generate_vbet_polygon(vbet_evidence_raster, rasterized_channel, channel_hand, out_valley_bottom, temp_folder, threshold=0.68):
 
-    with GeopackageLayer(gpkg_path, layer_name, write=True) as lyr:
-        lyr.create_layer(ogr_type, spatial_ref=srs)
-        for fname, ftype in attributes.items():
-            lyr.create_field(fname, ftype)
+    # Read initial rasters as arrays
+    vbet = raster2array(vbet_evidence_raster)
+    chan = raster2array(rasterized_channel)
+    hand = raster2array(channel_hand)
 
-        progbar = ProgressBar(len(list_geoms), 50, f"Saving {gpkg_path}/{layer_name}")
-        counter = 0
-        progbar.update(counter)
+    # Generate Valley Bottom
+    valley_bottom = ((vbet + chan) >= threshold) * ((hand + chan) > 0)  # ((A + B) < 0.68) * (C > 0)
+    valley_bottom_raw = os.path.join(temp_folder, "valley_bottom_raw.tif")
+    array2raster(valley_bottom_raw, vbet_evidence_raster, valley_bottom, data_type=gdal.GDT_Int32)
 
-        lyr.ogr_layer.StartTransaction()
-        for feat in list_geoms:
-            counter += 1
-            progbar.update(counter)
+    ds_valley_bottom = gdal.Open(valley_bottom_raw, gdal.GA_Update)
+    band_valley_bottom = ds_valley_bottom.GetRasterBand(1)
 
-            geom = feat['geom']
+    # Sieve and Clean Raster
+    gdal.SieveFilter(srcBand=band_valley_bottom, maskBand=None, dstBand=band_valley_bottom, threshold=10, connectedness=8, callback=gdal.TermProgress_nocb)
+    band_valley_bottom.SetNoDataValue(0)
+    band_valley_bottom.FlushCache()
+    valley_bottom_sieved = band_valley_bottom.ReadAsArray()
 
-            feature = ogr.Feature(lyr.ogr_layer_def)
-            geom_ogr = VectorBase.shapely2ogr(geom)
-            feature.SetGeometry(geom_ogr)
-            if attributes:
-                for field, value in feat['attributes'].items():
-                    feature.SetField(field, value)
-            lyr.ogr_layer.CreateFeature(feature)
-            feature = None
+    # Region Tool to find only connected areas
+    struct = generate_binary_structure(2, 2)
+    regions, _num = label(valley_bottom_sieved, structure=struct)
+    selection = regions * chan
+    values = np.unique(selection)
+    valley_bottom_region = np.isin(regions, values.nonzero())
+    array2raster(os.path.join(temp_folder, 'regions.tif'), vbet_evidence_raster, regions, data_type=gdal.GDT_Int32)
+    array2raster(os.path.join(temp_folder, 'valley_bottom_region.tif'), vbet_evidence_raster, valley_bottom_region.astype(int), data_type=gdal.GDT_Int32)
 
-        progbar.finish()
-        lyr.ogr_layer.CommitTransaction()
+    # Clean Raster Edges
+    valley_bottom_clean = binary_closing(valley_bottom_region.astype(int), iterations=2)
+    array2raster(out_valley_bottom, vbet_evidence_raster, valley_bottom_clean, data_type=gdal.GDT_Int32)
 
 
-def calculate_area(layer: Path, field_name: str, transform_epsg: int = 5070):
-    """Calcuate and store area (as hectares) for all features.
+def generate_centerline_surface(vbet_raster, out_cost_path, temp_folder):
 
-    Args:
-        layer (Path): path of layer to add area calculations
-        field_name (str): output field name. will overwrite data if field exists
-        transform_epsg (int, optional): epsg id of spatial reference (must be in meters!). Defaults to 5070 (Albers North America).
-    """
+    vbet = raster2array(vbet_raster)
 
-    with GeopackageLayer(layer, write=True) as lyr:
-        if lyr.ogr_layer.GetLayerDefn().GetFieldIndex(field_name) < 0:
-            lyr.create_field(field_name, ogr.OFTReal)
-        srs = lyr.get_srs_from_epsg(cfg.OUTPUT_EPSG)
-        _sr, transform = lyr.get_transform_from_epsg(srs, transform_epsg)
-        lyr.ogr_layer.StartTransaction()
-        for feat, *_ in lyr.iterate_features(f'Calculating area for features in {layer}'):
-            geom = feat.GetGeometryRef().Clone()
-            geom.Transform(transform)
-            area = geom.GetArea()
-            hectares = area / 10000.0
-            feat.SetField(field_name, hectares)
-            lyr.ogr_layer.SetFeature(feat)
-        lyr.ogr_layer.CommitTransaction()
+    # Generate Inverse Raster for Proximity
+    valley_bottom_inverse = (vbet != 1)
+    inverse_mask_raster = os.path.join(temp_folder, 'inverse_mask.tif')
+    array2raster(inverse_mask_raster, vbet_raster, valley_bottom_inverse)
+
+    # Proximity Raster
+    ds_valley_bottom_inverse = gdal.Open(inverse_mask_raster)
+    band_valley_bottom_inverse = ds_valley_bottom_inverse.GetRasterBand(1)
+    proximity_raster = os.path.join(temp_folder, 'proximity.tif')
+    _ds_proximity, band_proximity = new_raster(proximity_raster, vbet_raster, data_type=gdal.GDT_Int32)
+    gdal.ComputeProximity(band_valley_bottom_inverse, band_proximity, ['VALUES=1', "DISTUNITS=PIXEL", "COMPRESS=DEFLATE"])
+    band_proximity.SetNoDataValue(0)
+    band_proximity.FlushCache()
+    proximity = band_proximity.ReadAsArray()
+
+    # Rescale Raster
+    rescaled = np.interp(proximity, (proximity.min(), proximity.max()), (0.0, 10.0))
+    rescaled_raster = os.path.join(temp_folder, 'rescaled.tif')
+    array2raster(rescaled_raster, vbet_raster, rescaled, data_type=gdal.GDT_Float32)
+
+    # Centerline Cost Path
+    cost_path = 10**((rescaled * -1) + 10) + (rescaled <= 0) * 1000000000000  # 10** (((A) * -1) + 10) + (A <= 0) * 1000000000000
+    array2raster(out_cost_path, vbet_raster, cost_path, data_type=gdal.GDT_Float32)
 
 
 def create_project(huc, output_dir: str, meta: List[RSMeta], meta_dict: Dict[str, str]):
@@ -606,22 +461,29 @@ def create_project(huc, output_dir: str, meta: List[RSMeta], meta_dict: Dict[str
 
 
 def main():
-
     parser = argparse.ArgumentParser(
         description='Riverscapes VBET Tool',
         # epilog="This is an epilog"
     )
     parser.add_argument('huc', help='nhd watershed id', type=str)
     parser.add_argument('scenario_code', help='machine code for vbet scenario', type=str)
-    parser.add_argument('inputs', help='key-value pairs of input name and path', type=str)
-    parser.add_argument('vaa_table', type=str)
+    parser.add_argument('flowline_network', help='full nhd line network', type=str)
+    parser.add_argument('dem', help='dem', type=str)
+    parser.add_argument('slope', help='slope', type=str)
+    parser.add_argument('catchments', type=str)
+    parser.add_argument('channel_area', type=str)
+    parser.add_argument('vaa_table')
     parser.add_argument('output_dir', help='Folder where output VBET project will be created', type=str)
+    parser.add_argument('--level_paths', help='csv list of level paths', type=str, default="")
+    parser.add_argument('--pitfill', help='riverscapes project metadata as comma separated key=value pairs', default=None)
+    parser.add_argument('--dinfflowdir_ang', help='(optional) a little extra logging ', default=None)
+    parser.add_argument('--dinfflowdir_slp', help='Add debug tools for tracing things like memory usage at a performance cost.', default=None)
+    parser.add_argument('--twi_raster', help='Add debug tools for tracing things like memory usage at a performance cost.', default=None)
     parser.add_argument('--reach_codes', help='Comma delimited reach codes (FCode) to retain when filtering features. Omitting this option retains all features.', type=str)
     parser.add_argument('--flowline_type', type=str, default='NHD')
     parser.add_argument('--meta', help='riverscapes project metadata as comma separated key=value pairs', type=str)
     parser.add_argument('--verbose', help='(optional) a little extra logging ', action='store_true', default=False)
     parser.add_argument('--debug', help='Add debug tools for tracing things like memory usage at a performance cost.', action='store_true', default=False)
-
     args = dotenv.parse_args_env(parser)
 
     # make sure the output folder exists
@@ -633,24 +495,25 @@ def main():
     log.title('Riverscapes VBET For HUC: {}'.format(args.huc))
 
     meta = parse_metadata(args.meta)
-    inputs = parse_metadata(args.inputs)
-
     reach_codes = args.reach_codes.split(',') if args.reach_codes else None
+    level_paths = args.level_paths.split(',')
+    level_paths = level_paths if level_paths != ['.'] else None
 
-    try:
-        if args.debug is True:
-            from rscommons.debug import ThreadRun
-            memfile = os.path.join(args.output_dir, 'vbet_mem.log')
-            retcode, max_obj = ThreadRun(vbet, memfile, args.huc, args.scenario_code, inputs, args.vaa_table, args.output_dir, reach_codes, meta)
-            log.debug('Return code: {}, [Max process usage] {}'.format(retcode, max_obj))
+    # try:
+    if args.debug is True:
+        from rscommons.debug import ThreadRun
+        memfile = os.path.join(args.output_dir, 'vbet_mem.log')
+        retcode, max_obj = ThreadRun(vbet_centerlines, memfile, args.flowline_network, args.dem, args.slope, args.catchments, args.channel_area, args.vaa_table, args.output_dir, args.scenario_code, args.huc, args.pitfill, args.dinfflowdir_ang, args.dinfflowdir_slp, args.twi_raster, level_paths=level_paths, meta=meta)
+        log.debug('Return code: {}, [Max process usage] {}'.format(retcode, max_obj))
 
-        else:
-            vbet(args.huc, args.scenario_code, inputs, args.vaa_table, args.output_dir, reach_codes, meta, flowline_type=args.flowline_type)
+    else:
 
-    except Exception as e:
-        log.error(e)
-        traceback.print_exc(file=sys.stdout)
-        sys.exit(1)
+        vbet_centerlines(args.flowline_network, args.dem, args.slope, args.catchments, args.channel_area, args.vaa_table, args.output_dir, args.scenario_code, args.huc, level_paths, args.pitfill, args.dinfflowdir_ang, args.dinfflowdir_slp, args.twi_raster)
+
+    # except Exception as e:
+    #     log.error(e)
+    #     traceback.print_exc(file=sys.stdout)
+    #     sys.exit(1)
 
     sys.exit(0)
